requirements.txt
appdirs==1.4.4
asgiref==3.2.10
black==20.8b1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
Django==3.1.2
django-allauth==0.41.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.1
idna==2.10
Markdown==3.3.3
mypy-extensions==0.4.3
oauthlib==3.1.0
pathspec==0.8.0
Pillow==7.2.0
Pygments==2.7.2
python3-openid==3.2.0
pytz==2020.1
regex==2020.9.27
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
whitenoise==5.2.0
